+++ taskflowbackend/tests/test_flows.py
@@ -1018,6 +1018,42 @@ class TestLandingZoneMove(
         self.zone.refresh_from_db()
         self.assertEqual(self.zone.status, ZONE_STATUS_ACTIVE)
 
+    def test_validate_bom_header(self):
+        """Test landing_zone_move validation with BOM header in MD5 file"""
+        coll_path = os.path.join(self.zone_path, COLL_NAME)
+        zone_coll = self.irods.collections.create(coll_path)
+        obj = self.make_irods_object(zone_coll, OBJ_NAME)
+        obj_path = obj.path
+        # Make MD5 object with BOM header
+        md5_path = obj.path + '.md5'
+        md5_content = bytes(self.get_md5_checksum(obj), encoding='utf-8-sig')
+        make_object(self.irods, md5_path, md5_content)
+        self.assertEqual(self.irods.data_objects.exists(obj_path), True)
+        self.assertEqual(
+            self.irods.data_objects.exists(obj_path + '.md5'), True
+        )
+
+        flow_data = {
+            'zone_uuid': str(self.zone.sodar_uuid),
+            'validate_only': True,
+        }
+        flow = self.taskflow.get_flow(
+            irods_backend=self.irods_backend,
+            project=self.project,
+            flow_name='landing_zone_move',
+            flow_data=flow_data,
+        )
+        self.build_and_run(flow)
+
+        self.zone.refresh_from_db()
+        self.assertEqual(self.zone.status, ZONE_STATUS_ACTIVE)
+        self.assertEqual(self.irods.data_objects.exists(obj_path), True)
+        self.assertEqual(
+            self.irods.data_objects.exists(obj_path + '.md5'), True
+        )
+        sample_coll_path = os.path.join(self.sample_path, COLL_NAME)
+        self.assertEqual(self.irods.collections.exists(sample_coll_path), False)
+
     def test_validate_no_checksum(self):
         """Test landing_zone_move validation with missing checksum"""
         coll_path = os.path.join(self.zone_path, COLL_NAME)